requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
gunicorn==22.0.0
requests==2.32.2
python-dotenv==0.21.0
neo4j==6.0.2
google-genai==1.47.0
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
pandas==2.3.3
numpy==1.26.4
matplotlib==3.8.4
PyMuPDF==1.26.5
Pillow==10.3.0
pdfminer.six==20231228
python-docx==1.1.0
Flask-Cors==6.0.1
